+++ core/src/main/java/com/froso/ufp/modules/core/roles/controller/UserRightsController.java
@@ -14,51 +14,38 @@ import org.springframework.http.*;
 import org.springframework.stereotype.*;
 import org.springframework.web.bind.annotation.*;
 
-/**
- * Created with IntelliJ IDEA. SimpleUser:Christian Kleinhuis (ck@froso.de) Date: 16.11.13 Time: 20:57 To change
- * this
- * template use File | Settings | File Templates.
- */
 @Controller
-@RequestMapping("/" + UFPAuthenticateConstants.SESSION_PATH + "/" + RoleDefinition.TYPE_NAME)
+@RequestMapping("/" + UFPAuthenticateConstants.SESSION_PATH + "/" + UserRole.TYPE_NAME)
 @Api(description = BaseRepositoryController.CRUD_ADMIN_REPOSITORY,
-        tags = RoleDefinition.TYPE_NAME)
-class UserCapabilityController {
+        tags = UserRole.TYPE_NAME)
+class UserRightsController {
 
 
     @Autowired
-    private RoleCapabilityService roleCapabilityService;
+    private UserRightService roleRightService;
     @Autowired
-    private RoleDefinitionService roleDefinitionService;
+    private UserRoleService userRoleService;
     @Autowired
     private ICoreUserService coreUserService;
 
-    /**
-     * Gets all Elements from the repository
-     *
-     * @param userId  the user id
-     * @param request the request
-     * @return the elements
-     * @throws Exception the exception
-     */
     @ApiOperation(value = "get list of allowed user functions"
     )
     @RequestMapping(value = "",
             method = RequestMethod.GET)
     @ResponseBody
-    public ResponseEntity<BackendResponseTemplateSingleObject<RoleCapabilityResult>> readElements(
+    public ResponseEntity<BackendResponseTemplateSingleObject<RoleRightsResult>> readElements(
             @PathVariable("token") String userId) {
 
-        ResponseHandlerTemplateSingleObject<RoleCapabilityResult> manager = new ResponseHandlerTemplateSingleObject<>();
+        ResponseHandlerTemplateSingleObject<RoleRightsResult> manager = new ResponseHandlerTemplateSingleObject<>();
         // obtain current user
         AbstractCoreUser user = (AbstractCoreUser) coreUserService.findOne(userId);
-        RoleCapabilityResult userCapabilities = new RoleCapabilityResult();
+        RoleRightsResult userCapabilities = new RoleRightsResult();
         // iterate of user roles and collect all enabled functions
         for (DataDocumentLink role : user.getRoles()) {
-            RoleDefinition roleDef = roleDefinitionService.findOneBrute(role.getId());
+            UserRole roleDef = userRoleService.findOneBrute(role.getId());
             if (roleDef != null) {
                 for (DataDocumentLink capability : roleDef.getCapabilities()) {
-                    RoleCapability capDef = roleCapabilityService.findOneBrute(capability.getId());
+                    UserRight capDef = roleRightService.findOneBrute(capability.getId());
                     if (capDef != null) {
                         userCapabilities.add(capDef.getName().toLowerCase());
                     }