+++ core/src/main/java/com/froso/ufp/modules/core/roles/model/RoleRightsDefaultEnum.java
@@ -1,0 +1,11 @@
+package com.froso.ufp.modules.core.roles.model;
+
+import com.froso.ufp.modules.core.client.model.*;
+import com.froso.ufp.modules.core.resourcemetadata.annotations.*;
+import org.springframework.data.mongodb.core.mapping.*;
+
+public enum RoleRightsDefaultEnum {
+    admin    ,
+    user,
+    gues
+}