+++ core/src/main/java/com/froso/ufp/modules/core/user/model/ICoreUser.java
@@ -9,21 +9,9 @@ import java.util.*;
  * Created by ckleinhuis on 12.12.2016.
  */
 public interface ICoreUser extends IDataDocument {
-    /**
-     * Gets role.
-     *
-     * @return the role
-     */
-    UserRoleEnum getRole();
 
-    /**
-     * Sets role.
-     *
-     * @param role the role
-     */
-    void setRole(UserRoleEnum role);
 
-    Set<DataDocumentLink<RoleDefinition>> getRoles();
+    Set<DataDocumentLink<UserRole>> getRoles();
 
     String getFirstName();